+++ pybamm/models/submodels/interface/sei/__init__.py
@@ -3,3 +3,4 @@ from .total_sei import TotalSEI
 from .no_sei import NoSEI
 from .constant_sei import ConstantSEI
 from .sei_growth import SEIGrowth
+from .sei_thickness import SEIThickness