+++ pybamm/models/submodels/interface/sei/constant_sei.py
@@ -7,7 +7,7 @@ from .base_sei import BaseModel
 
 class ConstantSEI(BaseModel):
     """
-    Class for SEI with constant thickness.
+    Class for SEI with constant concentration.
 
     Note that there is no SEI current, so we don't need to update the "sum of
     interfacial current densities" variables from
@@ -31,11 +31,19 @@ class ConstantSEI(BaseModel):
             self.reaction_loc = "full electrode"
 
     def get_fundamental_variables(self):
+        # Constant concentrations
         domain = self.domain.lower()
-        # Constant thicknesses
-        L_inner = self.phase_param.L_inner_0
-        L_outer = self.phase_param.L_outer_0
-        variables = self._get_standard_thickness_variables(L_inner, L_outer)
+        if self.reaction_loc == "interface":
+            c_inner = self.phase_param.L_inner_0 / self.phase_param.V_bar_inner
+            c_outer = self.phase_param.L_outer_0 / self.phase_param.V_bar_outer
+        else:
+            c_inner = self.phase_param.a_typ * (
+                self.phase_param.L_inner_0 / self.phase_param.V_bar_inner
+            )
+            c_outer = self.phase_param.a_typ * (
+                self.phase_param.L_outer_0 / self.phase_param.V_bar_outer
+            )
+        variables = self._get_standard_concentration_variables(c_inner, c_outer)
 
         # Reactions
         if self.reaction_loc == "interface":
@@ -45,14 +53,5 @@ class ConstantSEI(BaseModel):
                 pybamm.Scalar(0), f"{domain} electrode", "current collector"
             )
         variables.update(self._get_standard_reaction_variables(zero, zero))
-
-        return variables
-
-    def get_coupled_variables(self, variables):
-        # Concentrations (derived from thicknesses)
-        variables.update(self._get_standard_concentration_variables(variables))
-
-        # Add other standard coupled variables
-        variables.update(super().get_coupled_variables(variables))
 
         return variables