+++ tests/integration/test_models/test_full_battery_models/test_lithium_ion/test_dfn.py
@@ -34,6 +34,43 @@ class TestDFN(BaseIntegrationTestLithiumIon, TestCase):
         param["Current function [A]"] = 0.5 * param["Nominal cell capacity [A.h]"]
         self.run_basic_processing_test({}, parameter_values=param)
 
+    def test_conservation_coupled_degradation(self):
+        # Test that lithium is conserved when all degradation mechanisms are enabled
+        model = pybamm.lithium_ion.DFN(
+            {
+                "SEI": "reaction limited",
+                "SEI porosity change": "true",
+                "lithium plating": "partially reversible",
+                "particle mechanics": ("swelling and cracking", "swelling only"),
+                "SEI on cracks": "true",
+                "loss of active material": "stress and reaction-driven",
+            }
+        )
+        param = pybamm.ParameterValues("OKane2022")
+
+        # optimize mesh size for mechanical degradation
+        var_pts = {"x_n": 5, "x_s": 5, "x_p": 5, "r_n": 30, "r_p": 20}
+        solver = pybamm.CasadiSolver(mode="fast")
+
+        # solve
+        sim = pybamm.Simulation(
+            model, parameter_values=param, var_pts=var_pts, solver=solver
+        )
+        solution = sim.solve([0, 3500])
+        BOD = solution["Total lithium in particles [mol]"].entries[0]
+        particles = solution["Total lithium in particles [mol]"].entries[-1]
+        neg = solution[
+            "Loss of lithium due to loss of active material in negative electrode [mol]"
+        ].entries[-1]
+        pos = solution[
+            "Loss of lithium due to loss of active material in positive electrode [mol]"
+        ].entries[-1]
+        side = solution["Total lithium lost to side reactions [mol]"].entries[-1]
+        EOD = particles + neg + pos + side
+
+        # compare
+        np.testing.assert_array_almost_equal(BOD, EOD, decimal=8)
+
 
 class TestDFNWithSizeDistribution(TestCase):
     def setUp(self):